Hide `iconv` functions on Apple devices behind a feature flag
